requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
